extendr-ffi: remove unnecessary `static mut` from bindings (#906)
* extendr-ffi: remove unnecessary `static mut` from bindings

* extendr-ffi: remove unnecessary `static mut` from bindings
Recent versions of Rust has started linting and posing warning on the use of `static mut`. Our bindings stems from `bindgen`, which does not have a mechanism to distinguish between `static`/`static mut` whenever it sees pointer.

Experience tells us the following: We are not to change the pointers of symbols nor environment variables. Therefore, these are no longer `static mut`, but `static`.

* Revert "extendr-ffi: remove unnecessary `static mut` from bindings"

This reverts commit 0bfa1de6c27fb063216acc1bc9a3acba69a7c244.

* ci: fixed the snapshot tests (for now)